requirements.txt
django-annoying>=0.10.4
django-markdown-app>=0.9.4.1
django-taggit>=0.22.2
pytz>=2018.5
django-hitcount>=1.3.0
django-uuslug==1.1.8
django-bootstrap3==10.0.1
django-debug-toolbar==1.10.1
django-widget-tweaks==1.4.3
django-avatar==4.1.0